Saved orhographic camera

+++ cpp/RT_CPP/RT_CPP/orthographic.cpp
@@ -2,7 +2,7 @@
 using namespace cameras;
 
 orthographic::orthographic(float farPlane, float focalLength, float fovRadians, vector3 upDirection, vector3 target, vector3 position,
-	float sceneWidth, float sceneHeight) :
+	int sceneWidth, int sceneHeight) :
 	camera(farPlane, focalLength, fovRadians, upDirection, target, position, sceneWidth, sceneHeight)
 {
